backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.25
alembic>=1.12.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
pandas>=2.2.0
numpy>=1.24.0
scikit-learn>=1.4.0
joblib>=1.3.0
faker>=20.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
python-jose>=3.3.0
cryptography>=41.0.0
passlib>=1.7.4
bcrypt>=4.0.0
